Add basic gallery component with modal
Gallery images that are clicked will open in a modal which can be
closed. Only the single image is shown within the modal for now.

Use a High Order Component for the gallery, which requires an image
store to be injected for shared state.

Note that constants are defined outside the render function (e.g.
galleryStore and GalleryImage) to avoid superfluous memory allocations
on every render.